+++ src/behavior/error.rs
@@ -16,6 +16,13 @@ use thiserror::Error;
 /// `behaviortree` behavior error type
 #[derive(Error, Debug)]
 pub enum BehaviorError {
+    
+    /// Pass through from nanoserde::DeJsonErr
+    #[error("{0}")]
+    Nanoserde(#[from] nanoserde::DeJsonErr),
+    /// Pass through from core::num::ParseIntError
+    #[error("{0}")]
+    ParseInt(#[from] core::num::ParseIntError),
     /// Pass through blackboard error
     #[error("{0}")]
     Blackboard(#[from] crate::blackboard::error::Error),