+++ src/port/mod.rs
@@ -48,6 +48,9 @@ pub fn is_bb_pointer(port: &str) -> bool {
 /// Remove blackboard pointer decorations from port name.
 #[must_use]
 pub fn strip_bb_pointer(port: &str) -> Option<ConstString> {
+    if !is_allowed_port_name(&port[1..]) {
+        return None
+    }
     Some(port.strip_prefix('{')?.strip_suffix('}')?.into())
 }